aosa-report/requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.12.0
urllib3==1.25.6
webencodings==0.5.1
PyGithub==1.44
